Attempted fix for a function scope issue.
Code like:
var ln = fnc();
...
function fnc() { w('here'); };

is ok but when wrapped like:
(function() {
  var ln = window['ln'] = fnc();
...
  var fnc = window['fnc'] = function() { w('here'); };
})();

would throw because fnc is not yet in scope

The attempted fix as outlined by richt@ is:
(function() {
  var ln = window['ln'] = fnc();
...
  function fnc() { w('here'); };
  var fnc = window['fnc'] = fnc;
})();

i.e. leaving the function declaration as it is but exporting it to
global scope just afterwards.