1 bug fix and 1 simplification:
It turns out that ast.Assign nodes would actually cause the lhs_finder
to fail (these are things like foo: bar in an object). I can't think of
any good reason to be looking at Assign nodes in the first place, so this
is removed (all tests still passing, and fixing the one failure).

Also, in lsh_finder, instead of splitting on "," then creating a list,
then iterating over each member of the list to find a match for lh_object,
just look for lh_object directly in the node string before splitting.