"Downgraded" the ElementTree syntax used for attributes.
The AOC servers run an older version of this library which doesn't
support the ele.find("[@foo]") syntax. The syntax is also supposed
to return a list of elements with this attribute, which seemed
unnecessary when accessing an attribute of root (widget) or
searching from a specific leaf element with the attr (content).